add super-resolution
