+++ README.md
@@ -1,31 +1,31 @@
 # StyleGAN2-Art [WIP]
 
-StyleGAN is a Generative Adversarial Network proposed by NVIDIA researchers. It builds upon the <a href="https://arxiv.org/pdf/1710.10196.pdf">Progressive Growing GAN</a> architecture to produce photo-realistic synthetic images. They have released three research papers regarding this as of June 2021. Their contributions are briefly summarized as follows:
+StyleGAN is a Generative Adversarial Network proposed by NVIDIA researchers. It builds upon the [Progressive Growing GAN](https://arxiv.org/pdf/1710.10196.pdf)architecture to produce photo-realistic synthetic images. They have released three research papers regarding this as of June 2021. Their contributions are briefly summarized as follows:
 
-- <a href="https://arxiv.org/pdf/1812.04948.pdf">StyleGAN </a><a href="./Architecture Diagrams/StyleGAN-Architecture.png">[Architecture]</a>:
-  - <b>Bilinear Sampling</b> to improve image quality.
-  - <b>Mapping Network</b> transforms latent space to w-space (Intermediate latent space).
-  - <b>Synthesis Network</b> uses <a href="https://arxiv.org/pdf/1703.06868.pdf">ADA-IN (Adaptive Instance Normalization)</a> to generate images.
-  - <b>Constant Initial Input</b> increases performance. W-space and ADA-IN control generated images anyway.
-  - <b>Gaussian Noise</b> makes the generated image look more realistic by bringing in finer features.
-  - <b>Mixing Regularization</b> performs style mixing where images are generated from two intermediate styles.
-  - <b>Perceptual Path Length</b> measures the difference between successive images when interpolating between two noise inputs.
-  - <b>Linear Separability</b> separates the latent space with a linear hyperplane based on attribute.
+- [StyleGAN](https://arxiv.org/pdf/1812.04948.pdf) [[Architecture](/diags/StyleGAN-Architecture.png)]:
+  - **Bilinear Sampling:** to improve image quality.
+  - **Mapping Network:** transforms latent space to w-space (Intermediate latent space).
+  - **Synthesis Network:** uses [ADA-IN (Adaptive Instance Normalization)](https://arxiv.org/pdf/1703.06868.pdf) to generate images.
+  - **Constant Initial Input:** increases performance. W-space and ADA-IN control generated images anyway.
+  - **Gaussian Noise:** makes the generated image look more realistic by bringing in finer features.
+  - **Mixing Regularization:** performs style mixing where images are generated from two intermediate styles.
+  - **Perceptual Path Length:** measures the difference between successive images when interpolating between two noise inputs.
+  - **Linear Separability:** separates the latent space with a linear hyperplane based on attribute.
 
-- <a href="https://arxiv.org/pdf/1912.04958.pdf">StyleGAN2 </a><a href="./Architecture Diagrams/StyleGAN2-Architecture.png">[Architecture]</a>:
-  - <b>Weight Demodulation</b> removes the droplet artifacts that were in the original StyleGAN.
-  - <b>Lazy Regularization</b> alleviates heavy memory usage and computation cost of regularization.
-  - <b>Perceptual Path Length Regularization</b> encourages smooth mapping from latent space to generated image to achieve feature disentanglement.
-  - <b>No Growing</b> replaces Progressive Growing GAN architecture to prevent phase artifacts with skip connections in the generator and residual connections in the discriminator.
-  - <b>Large Networks</b> yield better results where high-resolution images have more influence.
+- [StyleGAN2](https://arxiv.org/pdf/1912.04958.pdf) [[Architecture](/diags/StyleGAN2-Architecture.png)]:
+  - **Weight Demodulation:** removes the droplet artifacts that were in the original StyleGAN.
+  - **Lazy Regularization:** alleviates heavy memory usage and computation cost of regularization.
+  - **Perceptual Path Length Regularization:** encourages smooth mapping from latent space to generated image to achieve feature disentanglement.
+  - **No Growing:** replaces Progressive Growing GAN architecture to prevent phase artifacts with skip connections in the generator and residual connections in the discriminator.
+  - **Large Networks:** yield better results where high-resolution images have more influence.
 
-- <a href="https://arxiv.org/pdf/2006.06676.pdf">StyleGAN2-ADA </a><a href="./Architecture Diagrams/StyleGAN2-ADA-Architecture.png">[Architecture]</a>:
-  - <b>Adaptive Discrimator Augmentation</b> augments the data given to the discriminator to overcome overfitting without the augmentations leaking into generated images.
+- [StyleGAN2-ADA](https://arxiv.org/pdf/2006.06676.pdf) [[Architecture](/diags/StyleGAN2-ADA-Architecture.png)]:
+  - **Adaptive Discrimator Augmentation:** augments the data given to the discriminator to overcome overfitting without the augmentations leaking into generated images.
 
 
 ## Credits
 
-This repository is adapted from Phil Wang's <a href="https://github.com/lucidrains/stylegan2-pytorch">"Simple StyleGan2 for PyTorch"</a>. I found his version of the NVlabs official PyTorch implementation of <a href="https://github.com/NVlabs/stylegan2-ada-pytorch">"StyleGAN2-ADA"</a> to be easier to comprehend. His experiments are also interesting and informative.
+This repository is adapted from Phil Wang's [`lucidrains/stylegan2-pytorch`](https://github.com/lucidrains/stylegan2-pytorch). I found his version of the NVlabs official PyTorch implementation of [`NVlabs/stylegan2-ada-pytorch`](https://github.com/NVlabs/stylegan2-ada-pytorch) to be easier to comprehend. His experiments are also interesting and informative.
 
 
 ## Getting Started
@@ -34,16 +34,23 @@ This repository is adapted from Phil Wang's <a href="https://github.com/lucidrai
 
 Install "libsndfile" using the following commands for handling audio files in Linux:
 ```bash
-$ sudo apt update -y
-$ sudo apt install libsndfile1 -y
+sudo apt update -y
+sudo apt install libsndfile1 -y
 ```
-
 
 ### Install ffmpeg
 
 Install "ffmpeg" using the following commands:
 ```bash
-$ cd ~/ffmpeg_sources && \
+sudo apt install ffmpeg [Ubuntu]
+brew install ffmpeg     [macOS]
+```
+
+**Full install**
+<details>
+
+```
+cd ~/ffmpeg_sources && \
 wget -O ffmpeg-snapshot.tar.bz2 https://ffmpeg.org/releases/ffmpeg-snapshot.tar.bz2 && \
 tar xjvf ffmpeg-snapshot.tar.bz2 && \
 cd ffmpeg && \
@@ -75,6 +82,7 @@ make install && \
 hash -r
 ```
 
+</details>
 
 ### Install Packages
 
@@ -140,4 +148,4 @@ $ streamlit run run.py
 
 Check out the GIF below for demo!
 
-<img src="./app.gif">
+<img src="app.gif">
