+++ infer.py
@@ -1,8 +1,6 @@
-import os
-
 import torch
 
-from stylegan2.utils import audio_features, noise, image_noise, timestamped_filename
+from stylegan2.utils import audio_features, noise, timestamped_filename
 from stylegan2.model import StyleGAN2Model
 
 # file paths
@@ -27,4 +25,4 @@ noise_z = torch.randn(1, latent_dim).repeat(n, 1)
 noise = torch.FloatTensor(n, image_size, image_size, 1).uniform_(0., 1.)
 
 # generate images from uniform changes in latent space
-#model.generate_from_latent(name = timestamped_filename(), noise_z = noise_z, noise = noise)
+model.generate_from_latent(name = timestamped_filename(), noise_z = noise_z, noise = noise)
